client connection to act like server one
if i have to speak with alice, and alice may also want to connect with
me, we can end up creating two iroh connections in current code. this
change is an attempt to make this use a single connection if possible.

here we are calling process_connection on the client connection, which
allows the other side to not only handle our client requests, but also
send then client requests on the same connection.

the next step is to take server connections, created in
ftnet::server::run(), and it to the client bb8 pool, so if the other
side has already initiated a connection with us, we can reuse that
connection to send our requests to them too.