Fix: don't block on `fastn serve`
The [Identity::run] method blocked on `fastn serve` command and
never got to inserting the port in the [id_map].

We fix this by spawning the process instead of waiting for it to finish.
And, read its stdout until the point where we can parse port it's
running on.

Now the function only blocks until it has got the port from the child's
stdout.