not using bb8 connection pooling
this is an attempt to figure out why are things so unstable. this is
just a hypothesis, not backed by any real reason, other than trying to
simplify things.

the solution is not complete, we are creating connections, but not
cleaning them up on idle timeout, nor at connection errors etc.