Feat: `peer_to_peer` incoming request is now generic
Instead of taking `Request<hyper::body::Incoming>` we are now generic
over `Request<T>` This allows us to use `peer_to_peer` with both
`Request<Incoming>` and `Request<Bytes>`.

This is done because there seem to be no way of constructing our own
`Request<Incoming>`. This is needed if the request does not directly hit
the hyper service but instead comes from somewhere else (a test request
for example).