Doc: kulfi.app design details
The features/workings of the kulfi app is documented from the
perspective of various fastn apps that'll be run to:
- support Orgs
- support accounts and profiles

The current design does not factor in fine grained access controls with
the help of folder based permissions model supported by lets-auth. If
this was considered then account and profiled would be managed by two
separate fastn apps. For now, to keep things simple, Orgs are managed by
a fastn package and account+profiles are managed by another fastn
package.