Doc(org): The Dashboard Page journey
The Dashbaord Page explains multiple views and describes how it looks to
owner after adding a service and what all the owner can do.

It also links to /doc/journeys/add-service/ which will describe the
process of adding a service. And, /doc/journeys/join-request/ which will
describe the process of visitors making a request and owner
accpeting/rejecting them. This **does not exist** yet.