calling send.finish() in a better location
before this the second tcp connections used to forever weight, as
we never called finish, so the tcp connection is always open.

with this change, the connection is getting closed.

but this is not a complete fix for our tcp issue, the second
connection still takes 4seconds.

also our tcp connection handling code seems concurrent, so why is second
connection getting stuck? because we have no buffer on the connection
handler?

to be investigated.