Fix(kulfi/webview): follow redirects manually
The webview does not automatically follows redirects (301-399) but
instead shows a blank page.
The suspect is because it's not using a http(s) location but a kulfi://
url, the webview (webkit) is refusuing to follow the redirect. There's
no warning/error thrown in the dev tools console and I don't have
anything to prove this suspicion.

Manually handling the redirect response is working fine so I'm sticking
to that.