feat: Add streaming support in `http_to_peer`
Similar to peer_to_http, the http_to_peer function now sends the request
to the peer in a streaming fashion. It also takes the response and
streams it to the client.

Known Issues:

- kulfi does not compile yet. Only malai compiles
- The `LinesCodec` used by our FrameReader type is not helping us in
  streaming. It first reads the http headers in one line, then the
  entire body is sent as the next line. We need to move to a better
  handwritten encoding method where: 1. first 4 bytes are fixed length
  indicating the request header (first part) 2. subsequent chunk is of
  the said size that can be read as request header. 3. the rest of the
  stream is body content that should be streamed.

WARN: This commit breaks the compilation of the kulfi app but I have no
choice. The changes that I'm going to make to handle the "Known Issues"
will take some time and I'd like to push current work right now.