feat: Avoid Line encoding when reading response in http_to_peer
The line encoding tries to encode the body buffer as utf-8 and fails. We
avoid this by converting the FramedRead to RecvStream (from iroh) and
reading bytes from that.

For the json http response (headers and status), we read byte-wise in a
loop and construct the `kulfi_utils::http::Response` and then create a
Stream out of the RecvStream.

Ideally the LinesCodec should not be used at all as it creates problems
when trying to do HTTP streaming response but this small change solves
things.

NOTE: The kulfi side still does not compile yet.